feat(add): record button on teacherRecording component triggers text editor and audio recording
