+++ client/routes.js
@@ -5,6 +5,7 @@ import PropTypes from 'prop-types'
 import {Login, Signup, UserHome} from './components'
 import {me} from './store'
 
+
 /**
  * COMPONENT
  */
